fix test type error

+++ src/test/java/com/sggc/services/GameServiceTest.java
@@ -12,6 +12,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.boot.test.context.SpringBootTest;
 
 import java.io.IOException;
 import java.io.UncheckedIOException;
@@ -22,7 +23,6 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
-
 @ExtendWith(MockitoExtension.class)
 class GameServiceTest {
 
@@ -167,7 +167,7 @@ class GameServiceTest {
         exampleGame.setAppid(appid);
         exampleGame.setMultiplayer(multiplayer);
         exampleGame.setName(name);
-        exampleGame.setId(BigInteger.valueOf(10L));
+        exampleGame.setId(10);
         return exampleGame;
     }
